python/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloup==2.1.2
Cython==3.0.9
data==0.4
decorator==5.1.1
distro==1.9.0
elevenlabs==1.0.0b1
executing==2.0.1
fireworks-ai==0.13.0
funcsigs==1.0.2
future==1.0.0
glcontext==2.5.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
httpx-sse==0.4.0
idna==3.6
ipython==8.22.2
isosurfaces==0.1.0
jedi==0.19.1
latex==0.7.0
manim==0.18.0
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistralai==0.1.6
moderngl==5.10.0
moderngl-window==2.4.4
multipledispatch==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycairo==1.26.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyglet==2.0.14
Pygments==2.17.2
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
scipy==1.12.0
screeninfo==0.8.1
setuptools==69.2.0
shutilwhich==1.1.0
six==1.16.0
skia-pathops==0.8.0.post1
sniffio==1.3.1
srt==3.5.3
stack-data==0.6.3
svgelements==1.9.6
tempdir==0.7.1
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==3.0.0
wcwidth==0.2.13
websockets==12.0
